Updated svgs, Pr comment fixes
